Move ranges to edu.ida.la.ranges and exceptions to edu.ida.la.exceptions.
Extended the wrapper generators to generate "workspace-less" versions
of the functions when there exist functions which have arguments
like WORK, LWORK. These have the same name as the originals, but
take fewer arguments.

Now using syevr instead of syevd for symmetric eigenvectors. Seems to
be faster and syevd had problems converging on some of Gunnars
data :(


git-svn-id: https://ml01.zrz.tu-berlin.de/svn/ida/private/mikio/java-blas/trunk@3893 1449bcba-80d8-4576-a99e-0f548206d62b